refactor: move interfaces to usage place
