+++ internal/agent/reader/common.go
@@ -1,4 +1,4 @@
-package pkg
+package reader
 
 type MetricReader[T any] interface {
 	Get(name string) (T, bool)
